requirements.txt
coreapi>=2.3.3,<3  # Temporary. Remove if it's possible. There is dependency on drf-yasg, django-filter and djangorestframework
coreschema>=0.0.4,<1  # Temporary. Remove if it's possible. There is dependency on drf-yasg, django-filter and djangorestframework
Django>=4.2.2,<4.3
djangorestframework>=3.13.1,<4
psycopg2-binary>=2.9.9,<3
django-extensions>=3.1.5,<4
django-ckeditor>=6.7.0,<7
gunicorn[gevent]>=20.1.0,<21
drf-yasg>=1.20.0,<2
Pillow>=9.4.0,<10
django-debug-toolbar>=3.8.1,<4
ics>=0.7.2,<1
requests>=2.28.1,<3
django-admin-inline-paginator>=0.3.0,<1
awscli>=1.27.41,<2
boto3>=1.26.48,<2
django-storages>=1.13.2,<2
django-filter>=22.1,<23
django-password-validators==1.7.0
flake8>=4.0.1,<5
coverage>=6.3.2,<7
unittest-xml-reporting>=3.2.0,<4
python-jose>=3.3.0,<4
django-admin-rangefilter>=0.10.0,<1.0
maxminddb==2.3.0
pytz==2023.3
pycryptodome>=3.18.0, <4
